backend/requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2025.6.15
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.88.0
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.67.1
tzdata==2024.1
typing-extensions==4.14.0
typing_inspection==0.4.1
Werkzeug==3.0.3 